+++ kotest-assertions/kotest-assertions-shared/src/commonMain/kotlin/io/kotest/assertions/DescribeTypedMismatch.kt
@@ -1,0 +1,3 @@
+package io.kotest.assertions
+
+expect fun describeTypedMismatch(expected: Any?, actual: Any?): String
