+++ kotest-assertions/kotest-assertions-shared/src/desktopMain/kotlin/io/kotest/assertions/DescribeTypedMismatch.desktop.kt
@@ -1,0 +1,5 @@
+package io.kotest.assertions
+
+actual fun describeTypedMismatch(expected: Any?, actual: Any?): String {
+   TODO("Not yet implemented")
+}
