+++ kotest-assertions/kotest-assertions-shared/src/jsHostedMain/kotlin/io/kotest/assertions/DescribeTypedMismatch.jsHosted.kt
@@ -1,0 +1,3 @@
+package io.kotest.assertions
+
+actual fun describeTypedMismatch(expected: Any?, actual: Any?): String = ""
