+++ kotest-assertions/kotest-assertions-shared/src/nativeMain/kotlin/io/kotest/assertions/DescribeTypedMismatch.native.kt
@@ -1,0 +1,3 @@
+package io.kotest.assertions
+
+actual fun describeTypedMismatch(expected: Any?, actual: Any?): String = ""
